# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin Swagger API.",
        "title": "Gin Swagger Demo",
        "contact": {
            "name": "JLHou"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/CRUD": {
            "get": {
                "description": "列出所有",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRUD"
                ],
                "summary": "GetAllUser",
                "responses": {
                    "200": {
                        "description": "Get All Users Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "註冊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRUD"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Register account",
                        "name": "register_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Register Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/CRUD/login": {
            "post": {
                "description": "登入",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRUD"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login account",
                        "name": "login_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/CRUD/uA/{ID}": {
            "put": {
                "description": "更新使用者 Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRUD"
                ],
                "summary": "Update Account By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Account Struct",
                        "name": "update_account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UserRequestUpdateAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update User Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/CRUD/uP/{ID}": {
            "put": {
                "description": "更新使用者 Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRUD"
                ],
                "summary": "Update Password By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Password ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Password Struct",
                        "name": "update_password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UserRequestUpdatePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update User Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/CRUD/{id}": {
            "get": {
                "description": "列出欲查詢的使用者",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRUD"
                ],
                "summary": "GetUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get User Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "刪除使用者",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CRUD"
                ],
                "summary": "DeleteUserByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delete Account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete User Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Test": {
            "get": {
                "description": "測試在 GCP Cloud Run 是否連得上不帶 pSQL 的映像檔",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Connect"
                ],
                "summary": "Test Connect Cloud Run",
                "responses": {
                    "200": {
                        "description": "Connect Cloud Run Successful",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/c": {
            "post": {
                "description": "註冊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OLD"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Register account",
                        "name": "register_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/old.Customer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Register Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/d": {
            "delete": {
                "description": "刪除使用者",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OLD"
                ],
                "summary": "DeleteUserByAccount",
                "parameters": [
                    {
                        "description": "Delete user",
                        "name": "delete_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/old.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete User Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "登入",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OLD"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login account",
                        "name": "login_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/old.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/r": {
            "get": {
                "description": "列出所有",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OLD"
                ],
                "summary": "GetUsers",
                "responses": {
                    "200": {
                        "description": "Get Users Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/test_connect": {
            "get": {
                "description": "測試在 GCP Cloud Run 是否連得上不帶 pSQL 的映像檔",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OLD"
                ],
                "summary": "Connect Test",
                "responses": {
                    "200": {
                        "description": "Connect Cloud Run Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ua/{ID}": {
            "put": {
                "description": "更新Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OLD"
                ],
                "summary": "Update Account By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Account Struct",
                        "name": "update_account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/old.UpdateAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update User Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/up/{ID}": {
            "put": {
                "description": "更新Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OLD"
                ],
                "summary": "Update Password By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Password ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Password Struct",
                        "name": "update_password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/old.UpdatePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update User Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.UserRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "description": "帳號",
                    "type": "string"
                },
                "password": {
                    "description": "密碼",
                    "type": "string"
                }
            }
        },
        "http.UserRequestUpdateAccount": {
            "type": "object",
            "properties": {
                "new_account": {
                    "description": "更新帳號",
                    "type": "string"
                }
            }
        },
        "http.UserRequestUpdatePassword": {
            "type": "object",
            "properties": {
                "new_password": {
                    "description": "更新密碼",
                    "type": "string"
                }
            }
        },
        "old.Customer": {
            "type": "object",
            "properties": {
                "cust_account": {
                    "description": "帳號",
                    "type": "string"
                },
                "cust_password": {
                    "description": "密碼",
                    "type": "string"
                }
            }
        },
        "old.UpdateAccount": {
            "type": "object",
            "properties": {
                "NewAccount": {
                    "description": "欲更新的帳號",
                    "type": "string"
                }
            }
        },
        "old.UpdatePassword": {
            "type": "object",
            "properties": {
                "NewPassword": {
                    "description": "欲更新的密碼",
                    "type": "string"
                }
            }
        }
    }
}